+++ CHANGES.md
@@ -8,12 +8,18 @@ Change Log
   * Removed `Camera.tilt`, which was deprecated in Cesium 1.6. Use `Camera.pitch`.
   * Removed `Camera.heading` and `Camera.tilt`. They were deprecated in Cesium 1.6. Use `Camera.setView`.
   * Removed `Camera.setPositionCartographic`, which was was deprecated in Cesium 1.6. Use `Camera.setView`.
+  * The `InfoBox` widget no longer sanitizes HTML descriptions by default. To go back to the old behavior, set `Cesium.InfoBoxViewModel.defaultPreprocessor` to `Cesium.sanitizeHtml` at the beginning of your application.
+* Deprecated
+  * `InfoBoxViewModel.defaultSanitizer` has been deprecated and will be removed in 1.10.  Use `InfoBoxViewModel.defaultPreprocessor` instead.
+  * `InfoBoxViewModel.sanitizer` has been deprecated and will be removed in 1.10.  Use `InfoBoxViewModel.preprocessor` instead.
+  * `InfoBoxViewModel.descriptionRawHtml` has been deprecated and will be removed in 1.10.  Use `InfoBoxViewModel.description` instead.
 * Fixed incorrect ellipse texture coordinates. [#2363](https://github.com/AnalyticalGraphicsInc/cesium/issues/2363) and [#2465](https://github.com/AnalyticalGraphicsInc/cesium/issues/2465)
 * Fixed a bug in imagery loading that could cause some or all of the globe to be missing when using an imagery layer that does not cover the entire globe.
 * Added support for rendering a water effect on Quantized-Mesh terrain tiles.
 * Added `pack` and `unpack` functions to `Matrix2` and `Matrix3`.
 * Added camera-terrain collision detection/response when the camera reference frame is set.
 * Added `ScreenSpaceCameraController.enableCollisionDetection` to enable/disable camera collision detection with terrain.
+* Added `sanitizeHtml`, which uses [Google Caja](https://developers.google.com/caja/) to asynchronously sanitize HTML.
 
 ### 1.6 - 2015-02-02